+++ crates/cosmos/cosmos-relayer/src/impls/relay/error.rs
@@ -1,10 +1,10 @@
 use eyre::eyre;
 use hermes_relayer_components::chain::traits::types::chain_id::HasChainId;
-use hermes_relayer_components::relay::components::create_client::CanRaiseMissingCreateClientEventError;
 use hermes_relayer_components::relay::impls::channel::open_init::CanRaiseMissingChannelInitEventError;
 use hermes_relayer_components::relay::impls::channel::open_try::CanRaiseMissingChannelTryEventError;
 use hermes_relayer_components::relay::impls::connection::open_init::CanRaiseMissingConnectionInitEventError;
 use hermes_relayer_components::relay::impls::connection::open_try::CanRaiseMissingConnectionTryEventError;
+use hermes_relayer_components::relay::impls::create_client::CanRaiseMissingCreateClientEventError;
 use hermes_relayer_components::relay::traits::target::{DestinationTarget, SourceTarget};
 use hermes_relayer_components_extra::relay::components::packet_relayers::retry::SupportsPacketRetry;
 use ibc_relayer_types::core::ics24_host::identifier::{ChannelId, ConnectionId};
