+++ crates/relayer/relayer-components-extra/src/components/extra/relay.rs
@@ -1,12 +1,12 @@
 use cgp_core::prelude::*;
 use cgp_core::RunnerComponent;
 use hermes_relayer_components::components::default::relay::DefaultRelayComponents;
-use hermes_relayer_components::relay::components::message_senders::chain_sender::SendIbcMessagesToChain;
-use hermes_relayer_components::relay::components::message_senders::update_client::SendIbcMessagesWithUpdateClient;
-use hermes_relayer_components::relay::components::packet_relayers::general::filter_relayer::FilterRelayer;
-use hermes_relayer_components::relay::components::packet_relayers::general::full_relay::FullCycleRelayer;
-use hermes_relayer_components::relay::components::packet_relayers::general::lock::LockPacketRelayer;
-use hermes_relayer_components::relay::components::packet_relayers::general::log::LoggerRelayer;
+use hermes_relayer_components::relay::impls::message_senders::chain_sender::SendIbcMessagesToChain;
+use hermes_relayer_components::relay::impls::message_senders::update_client::SendIbcMessagesWithUpdateClient;
+use hermes_relayer_components::relay::impls::packet_relayers::general::filter_relayer::FilterRelayer;
+use hermes_relayer_components::relay::impls::packet_relayers::general::full_relay::FullCycleRelayer;
+use hermes_relayer_components::relay::impls::packet_relayers::general::lock::LockPacketRelayer;
+use hermes_relayer_components::relay::impls::packet_relayers::general::log::LoggerRelayer;
 use hermes_relayer_components::relay::traits::channel::open_ack::ChannelOpenAckRelayerComponent;
 use hermes_relayer_components::relay::traits::channel::open_confirm::ChannelOpenConfirmRelayerComponent;
 use hermes_relayer_components::relay::traits::channel::open_handshake::ChannelOpenHandshakeRelayerComponent;
