+++ crates/relayer/relayer-components/src/components/default/relay.rs
@@ -1,24 +1,8 @@
 use cgp_core::prelude::*;
 use cgp_core::RunnerComponent;
 
-use crate::relay::components::auto_relayers::both_targets::RelayBothTargets;
-use crate::relay::components::auto_relayers::event::RelayEvents;
-use crate::relay::components::create_client::CreateClientWithChains;
-use crate::relay::components::event_relayers::packet_event::PacketEventRelayer;
-use crate::relay::components::message_senders::chain_sender::SendIbcMessagesToChain;
-use crate::relay::components::message_senders::update_client::SendIbcMessagesWithUpdateClient;
-use crate::relay::components::packet_clearers::receive_packet::ClearReceivePackets;
-use crate::relay::components::packet_relayers::ack::base_ack_packet::BaseAckPacketRelayer;
-use crate::relay::components::packet_relayers::general::filter_relayer::FilterRelayer;
-use crate::relay::components::packet_relayers::general::full_relay::FullCycleRelayer;
-use crate::relay::components::packet_relayers::general::lock::LockPacketRelayer;
-use crate::relay::components::packet_relayers::general::log::LoggerRelayer;
-use crate::relay::components::packet_relayers::receive::base_receive_packet::BaseReceivePacketRelayer;
-use crate::relay::components::packet_relayers::receive::skip_received_packet::SkipReceivedPacketRelayer;
-use crate::relay::components::packet_relayers::timeout_unordered::timeout_unordered_packet::BaseTimeoutUnorderedPacketRelayer;
-use crate::relay::components::update_client::build::BuildUpdateClientMessages;
-use crate::relay::components::update_client::skip::SkipUpdateClient;
-use crate::relay::components::update_client::wait::WaitUpdateClient;
+use crate::relay::impls::auto_relayers::both_targets::RelayBothTargets;
+use crate::relay::impls::auto_relayers::event::RelayEvents;
 use crate::relay::impls::channel::open_ack::RelayChannelOpenAck;
 use crate::relay::impls::channel::open_confirm::RelayChannelOpenConfirm;
 use crate::relay::impls::channel::open_handshake::RelayChannelOpenHandshake;
@@ -29,6 +13,22 @@ use crate::relay::impls::connection::open_confirm::RelayConnectionOpenConfirm;
 use crate::relay::impls::connection::open_handshake::RelayConnectionOpenHandshake;
 use crate::relay::impls::connection::open_init::InitializeConnection;
 use crate::relay::impls::connection::open_try::RelayConnectionOpenTry;
+use crate::relay::impls::create_client::CreateClientWithChains;
+use crate::relay::impls::event_relayers::packet_event::PacketEventRelayer;
+use crate::relay::impls::message_senders::chain_sender::SendIbcMessagesToChain;
+use crate::relay::impls::message_senders::update_client::SendIbcMessagesWithUpdateClient;
+use crate::relay::impls::packet_clearers::receive_packet::ClearReceivePackets;
+use crate::relay::impls::packet_relayers::ack::base_ack_packet::BaseAckPacketRelayer;
+use crate::relay::impls::packet_relayers::general::filter_relayer::FilterRelayer;
+use crate::relay::impls::packet_relayers::general::full_relay::FullCycleRelayer;
+use crate::relay::impls::packet_relayers::general::lock::LockPacketRelayer;
+use crate::relay::impls::packet_relayers::general::log::LoggerRelayer;
+use crate::relay::impls::packet_relayers::receive::base_receive_packet::BaseReceivePacketRelayer;
+use crate::relay::impls::packet_relayers::receive::skip_received_packet::SkipReceivedPacketRelayer;
+use crate::relay::impls::packet_relayers::timeout_unordered::timeout_unordered_packet::BaseTimeoutUnorderedPacketRelayer;
+use crate::relay::impls::update_client::build::BuildUpdateClientMessages;
+use crate::relay::impls::update_client::skip::SkipUpdateClient;
+use crate::relay::impls::update_client::wait::WaitUpdateClient;
 use crate::relay::traits::channel::open_ack::ChannelOpenAckRelayerComponent;
 use crate::relay::traits::channel::open_confirm::ChannelOpenConfirmRelayerComponent;
 use crate::relay::traits::channel::open_handshake::ChannelOpenHandshakeRelayerComponent;
