--- crates/relayer/relayer-components/src/relay/components/mod.rs
@@ -1,8 +1,0 @@
-pub mod auto_relayers;
-pub mod create_client;
-pub mod event_relayers;
-pub mod message_senders;
-pub mod packet_clearers;
-pub mod packet_filters;
-pub mod packet_relayers;
-pub mod update_client;
