+++ crates/relayer/relayer-components/src/relay/impls/mod.rs
@@ -1,2 +1,10 @@
+pub mod auto_relayers;
 pub mod channel;
 pub mod connection;
+pub mod create_client;
+pub mod event_relayers;
+pub mod message_senders;
+pub mod packet_clearers;
+pub mod packet_filters;
+pub mod packet_relayers;
+pub mod update_client;
