+++ crates/relayer/relayer-components/src/relay/mod.rs
@@ -27,7 +27,6 @@
    chains.
 */
 
-pub mod components;
 pub mod impls;
 pub mod traits;
 pub mod types;
